refactor: install sql dependencies for regactor to postgres
