+++ package.json
@@ -31,7 +31,9 @@
     "express": "^4.15.0",
     "jest": "^23.6.0",
     "jquery": "^3.1.1",
+    "knex": "^0.16.3",
     "mongoose": "^4.9.6",
+    "pg": "^7.8.0",
     "react-awesome-modal": "^2.0.5",
     "react-test-renderer": "^15.6.2",
     "request": "^2.81.0",
